Update ngx_http_vhost_traffic_status_module.c
Prepare code for second 32bit variable which deals with an overflow in stats-in/out bytes, let me know what you think, may have over done it if overflow detection is good in one place, but I found 2 places where in/out is incremented. If you agree or amend the code I can work on the javascript side to add (in = in + in_ov*4.2gb).
A local test shows json (":{"localhost":{"requestCounter":3,"inBytes":1113,"inBytesOv":0,"outBytes":6263,"outBytesOv":0,)
so it looks like it works in principle.